requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiosqlite==0.17.0
alembic==1.6.5
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
databases==0.4.3
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
multidict==5.1.0
pycparser==2.20
PyMySQL==0.9.3
pyTelegramBotAPI==3.8.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.3.24
typing-extensions==3.10.0.0
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
yarl==1.6.3
